rename: 채팅 도메인 v1 디렉토리 생성 및 분리
