+++ chat/src/main/java/com/team15gijo/chat/application/dto/v1/ChatRoomResponseDto.java
@@ -1,7 +1,7 @@
 package com.team15gijo.chat.application.dto.v1;
 
-import com.team15gijo.chat.domain.model.ChatRoomParticipant;
-import com.team15gijo.chat.domain.model.ChatRoomType;
+import com.team15gijo.chat.domain.model.v1.ChatRoomParticipant;
+import com.team15gijo.chat.domain.model.v1.ChatRoomType;
 import java.util.Set;
 import java.util.UUID;
 import lombok.Builder;
