+++ chat/src/main/java/com/team15gijo/chat/domain/repository/v1/ChatRoomParticipantRepository.java
@@ -1,6 +1,6 @@
 package com.team15gijo.chat.domain.repository.v1;
 
-import com.team15gijo.chat.domain.model.ChatRoomParticipant;
+import com.team15gijo.chat.domain.model.v1.ChatRoomParticipant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomParticipantRepository extends JpaRepository<ChatRoomParticipant, Long> {
