+++ chat/src/main/java/com/team15gijo/chat/domain/repository/v1/ChatRoomRepository.java
@@ -1,6 +1,6 @@
 package com.team15gijo.chat.domain.repository.v1;
 
-import com.team15gijo.chat.domain.model.ChatRoom;
+import com.team15gijo.chat.domain.model.v1.ChatRoom;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
