+++ chat/src/main/java/com/team15gijo/chat/presentation/dto/v1/ChatMessageResponseDto.java
@@ -1,6 +1,6 @@
 package com.team15gijo.chat.presentation.dto.v1;
 
-import com.team15gijo.chat.domain.model.ConnectionType;
+import com.team15gijo.chat.domain.model.v1.ConnectionType;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
